# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 経理・会計オールインワンAIアプリケーション API
  version: "1.0.0"
  description: |
    中小企業〜中堅企業向け「経理・会計をこのアプリ1つで完結させる」オールインワンAIプロダクトのAPI仕様。

    対応する統合DB設計は `docs/03_database_design.md` / `sql/001_initial_schema_all_in_one.sql` を参照。

    ## 設計原則
    - AIは提案(`/ai/suggestions`)のみを行い、確定処理(仕訳確定・税額計算等)は本APIの決定的ロジックが行う。
    - 状態遷移・業務アクションはサブリソース(例: `POST /journal-entries/{id}/post`)として明示する。
    - 全エンドポイントはマルチテナントであり、`X-Tenant-ID` ヘッダーとJWTクレームの双方で検証される。
    - 副作用を伴うPOST操作は `Idempotency-Key` ヘッダーによる冪等性をサポートする。
  contact:
    name: 経理会計オールインワンソフト開発チーム
servers:
  - url: https://api.keiri-kaikei.example.com/v1
    description: 本番環境
  - url: https://api.staging.keiri-kaikei.example.com/v1
    description: ステージング環境
  - url: http://localhost:3000/v1
    description: ローカル開発環境
...
security:
  - BearerAuth: []

# ==============================================================================
# PATHS
# ==============================================================================
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: メールアドレス・パスワードでログインする
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: ログイン成功。MFAが有効な場合は `mfa_required=true` とともに `mfa_token` を返す。
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: object
                    properties:
                      access_token: { type: string }
                      refresh_token: { type: string }
                      mfa_required: { type: boolean }
                      mfa_token: { type: string, description: "mfa_required=trueの場合のみ" }
                      tenants:
                        type: array
                        description: 所属テナント一覧(複数所属時はテナント選択が必要)
                        items:
                          type: object
                          properties:
                            tenant_id: { type: string, format: uuid }
                            tenant_name: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /auth/mfa:
    post:
      tags: [Auth]
      summary: MFA(多要素認証)コードを検証しアクセストークンを発行する
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfa_token, code]
              properties:
                mfa_token: { type: string }
                code: { type: string, description: "6桁のTOTPコード" }
      responses:
        "200":
          description: MFA検証成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: object
                    properties:
                      access_token: { type: string }
                      refresh_token: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: ログアウトし現在のリフレッシュトークンを失効させる
      responses:
        "204": { description: ログアウト成功 }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/signup:
    post:
      tags: [Auth]
      summary: 初回サインアップ(新規テナント作成 + オーナーユーザー登録)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupRequest" }
      responses:
        "201":
          description: サインアップ成功。作成されたテナントのオーナーとしてセッションを発行する
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: "#/components/schemas/AuthSession" }
                  meta: { $ref: "#/components/schemas/Meta" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /auth/invitations/validate:
    get:
      tags: [Auth]
      summary: 招待トークンの有効性を検証する
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 有効な招待。招待先のテナント名・メールアドレス・ロールを返す
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: "#/components/schemas/InvitationValidation" }
                  meta: { $ref: "#/components/schemas/Meta" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /auth/accept-invite:
    post:
      tags: [Auth]
      summary: 招待を受諾しテナントへ参加する(新規ユーザー作成、または既存ユーザーへの紐付け)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcceptInviteRequest" }
      responses:
        "200":
          description: 招待受諾成功。参加したテナントのセッションを発行する
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: "#/components/schemas/AuthSession" }
                  meta: { $ref: "#/components/schemas/Meta" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }

  # ----------------------------------------------------------------------------
  # Tenants / Users
  # ----------------------------------------------------------------------------
...
  /attachments:
    get:
      tags: [Attachments]
      summary: "添付ファイルを検索する(電帳法検索要件: 取引年月日/金額/取引先)"
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: transaction_date_from
          in: query
          schema: { type: string, format: date }
        - name: transaction_date_to
          in: query
          schema: { type: string, format: date }
        - name: amount
          in: query
          schema: { type: number }
        - name: counterparty_name
          in: query
          schema: { type: string }
          description: 部分一致検索(pg_trgm)
      responses:
        "200":
          description: 添付ファイル一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentListResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Attachments]
      summary: 証憑ファイルをアップロードする
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, transaction_date, amount, counterparty_name]
              properties:
                file: { type: string, format: binary }
                transaction_date: { type: string, format: date }
                amount: { type: number }
                counterparty_name: { type: string }
      responses:
        "201":
          description: アップロード成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /attachments/{id}:
    get:
      tags: [Attachments]
      summary: 添付ファイルのメタデータを取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 添付ファイル詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentResponse"
        "404": { $ref: "#/components/responses/NotFound" }
  /attachments/{id}/links:
    post:
      tags: [Attachments]
      summary: 添付ファイルを業務レコード(仕訳/経費明細/請求書等)へ関連付ける
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [linkable_type, linkable_id]
              properties:
                linkable_type:
                  type: string
                  enum: [journal_entry, expense_report_line, invoice, vendor_bill, fixed_asset]
                linkable_id: { type: string, format: uuid }
      responses:
        "201": { description: 関連付け成功 }
        "409": { $ref: "#/components/responses/Conflict" }

  # ----------------------------------------------------------------------------
  # Audit Logs (追記専用)
  # ----------------------------------------------------------------------------
...
  /auto-journal-rules:
    get:
      tags: [AutoJournalRules]
      summary: 自動仕訳ルール一覧を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: 自動仕訳ルール一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutoJournalRuleListResponse" }
    post:
      tags: [AutoJournalRules]
      summary: 自動仕訳ルールを作成する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutoJournalRuleCreate" }
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutoJournalRuleResponse" }
  /auto-journal-rules/{id}:
    patch:
      tags: [AutoJournalRules]
      summary: 自動仕訳ルールを更新する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutoJournalRuleCreate" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutoJournalRuleResponse" }
    delete:
      tags: [AutoJournalRules]
      summary: 自動仕訳ルールを無効化する(is_active=false。物理削除は行わない)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "204": { description: 無効化成功 }

  # ----------------------------------------------------------------------------
  # Customers / Vendors (取引先マスタ)
  # ----------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWTアクセストークン。`tenant_id` クレームを含み、`X-Tenant-ID` ヘッダーと
        一致しない場合はAPI層で403 (`TENANT_MISMATCH`) を返す。
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: 操作対象テナントのUUID。JWTクレームの tenant_id と一致する必要がある。
      schema: { type: string, format: uuid }
...
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    BadRequest:
      description: リクエストが不正です
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: 認証に失敗しました(トークン欠落・無効・期限切れ)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFound:
      description: リソースが見つかりません(RLSにより越境アクセス時も同一の404を返す)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: 状態遷移ルール・DB制約違反等によりリクエストを実行できません
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnprocessableEntity:
      description: 入力値のバリデーションエラー
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    Meta:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        pagination:
          type: object
          properties:
            page: { type: integer }
            page_size: { type: integer }
            total_count: { type: integer }
            total_pages: { type: integer }
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            例: INVALID_BALANCE, VALIDATION_ERROR, NOT_FOUND, UNAUTHORIZED, FORBIDDEN,
            TENANT_MISMATCH, IDEMPOTENCY_KEY_CONFLICT, VOID_WINDOW_EXPIRED,
            SELF_APPROVAL_NOT_ALLOWED, DUPLICATE_IMPORT, CONFLICT, ALREADY_EXISTS,
            INVITATION_EXPIRED, INVITATION_REVOKED, INVITATION_ALREADY_ACCEPTED
        message: { type: string }
        details:
          type: array
          items: { type: object }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: "#/components/schemas/ApiError" }

    # ---- Enums (DBのENUM型と対応) ----
...
    SignupRequest:
      type: object
      required: [email, password, name, tenant_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        name: { type: string, description: "ユーザー名" }
        tenant_name: { type: string, description: "会社/テナント名" }
    AuthSession:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        mfa_required: { const: false }
        tenants:
          type: array
          items:
            type: object
            properties:
              tenant_id: { type: string, format: uuid }
              tenant_name: { type: string }
    InvitationValidation:
      type: object
      properties:
        tenant_name: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [owner, accounting_manager, accountant, approver, employee, payroll_admin, viewer_external, system_service]
    AcceptInviteRequest:
      type: object
      required: [token, password, name]
      properties:
        token: { type: string, description: "招待メール/URLに含まれる生トークン" }
        password: { type: string, format: password, minLength: 8 }
        name: { type: string, description: "ユーザー名(既存アカウントに紐付く場合は無視される)" }

    # ---- Tenant / User ----
...
    Attachment:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        file_name: { type: string }
        storage_path: { type: string, readOnly: true }
        mime_type: { type: string }
        file_hash: { type: string, readOnly: true }
        transaction_date: { type: string, format: date }
        amount: { type: number }
        counterparty_name: { type: string }
        uploaded_by: { type: string, format: uuid, readOnly: true }
        uploaded_at: { type: string, format: date-time, readOnly: true }
    AttachmentResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/Attachment" }
        meta: { $ref: "#/components/schemas/Meta" }
    AttachmentListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Audit Logs ----
...
    AutoJournalRule:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        rule_name: { type: string }
        priority: { type: integer, default: 100 }
        source: { type: string, enum: [bank, card] }
        match_pattern: { type: string }
        min_amount: { type: number, nullable: true }
        max_amount: { type: number, nullable: true }
        debit_account_id: { type: string, format: uuid, nullable: true }
        credit_account_id: { type: string, format: uuid, nullable: true }
        is_active: { type: boolean, default: true }
    AutoJournalRuleCreate:
      type: object
      required: [rule_name, source, match_pattern]
      properties:
        rule_name: { type: string }
        priority: { type: integer, default: 100 }
        source: { type: string, enum: [bank, card] }
        match_pattern: { type: string }
        min_amount: { type: number }
        max_amount: { type: number }
        debit_account_id: { type: string, format: uuid }
        credit_account_id: { type: string, format: uuid }
    AutoJournalRuleResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/AutoJournalRule" }
        meta: { $ref: "#/components/schemas/Meta" }
    AutoJournalRuleListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/AutoJournalRule" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Customers / Vendors ----
...